services/face-recognition/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
insightface==0.7.3
onnxruntime==1.16.3
opencv-python-headless==4.9.0.80
numpy==1.26.3
httpx==0.26.0
pyyaml==6.0.1
python-multipart==0.0.6
pillow==10.2.0
